feat: add OptimizationProblem class for defining optimization problems
